카메라 컴포넌트 추가 중

+++ Source/TGON/Graphics/Render/RenderStage.h
@@ -23,7 +23,7 @@ public:
 
 /* @section Public type */
 public:
-    using CameraHandle = int32_t;
+    using CameraHandle = size_t;
 
 /* @section Public method */
 public:
